refactor: ♻️ retype campaign event consumers from lib/campaign/events
`TeamContractEventList.vue` and its spec now pull
`EventsByCampaignCode` / `ExtendedEvent` from the new pure module.
Also drop the stale `vi.mock('@/services/AddCampaignService', …)`
blocks in `TeamContractDetail*.spec.ts` — the component under test
never imported the service.

Refs #977
